Fix: Resolve metadata export error and remove Clarity
- Moved metadata export from `app/layout.tsx` to `app/metadata.ts` to resolve 'use client' conflict.
- Removed all references to Microsoft Clarity, including the component and mentions in the cookies policy.
