+++ app/metadata.ts
@@ -1,0 +1,19 @@
+import type { Metadata } from "next";
+
+export const metadata: Metadata = {
+  title: "Next Generation of Analytics | AnalyticsFlow",
+  description: "Data Analytical service with focus on making companies better",
+  openGraph: {
+    title: "Next Generation of Analytics | AnalyticsFlow",
+    description: "Data Analytical service with focus on making companies better",
+    url: "https://analyticsflow.cz",
+    type: "website",
+    images: [
+      {
+        url: "/data_main_photo.jpeg",
+      },
+    ],
+    siteName: "AnalyticsFlow",
+    locale: "en_US",
+  },
+};
